refactor: 도메인을 DTO로 변환하는 역할을 하는 클래스 이름을 Assembler로 변경, depth를 컨벤션에 맞게 수정
